requirements.txt
# Bot Gateway
python-telegram-bot==21.6

# Embedding Service (local model)
sentence-transformers>=3.0.0
numpy>=1.26.0

# RAG Service (OpenAI for LLM generation only)
openai>=1.50.0

# Vision Service (local BLIP model from HuggingFace)
transformers>=4.40.0
torch>=2.0.0
Pillow>=10.4.0

# Shared
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
aiosqlite>=0.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
